First attempt: sliders to dictate range, grey boxes in separate canvas, no conversions.
The markers have now been added to a canvas to the right of the main canvas.
The two arrows (top and bottom ranges) scroll throughout the entire range.
Refocus filters the data to the specified range.
There are issues with the fact that we are overlaying a canvas.
We need to include time conversions in order to go from absolute, relative,
and screen locations. This is a WIP